+++ app/windows/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   app_links
   emoji_picker_flutter
   file_selector_windows
+  firebase_core
   flutter_secure_storage_windows
   record_windows
   share_plus